+++ back/src/main/java/com/back/auth/domain/OAuthAccountRepository.java
@@ -5,11 +5,11 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /** OAuth 계정 매핑 조회/검증용 JPA 리포지토리. */
-public interface OAuthAccountRepository extends JpaRepository<OAuthAccount, Integer> {
+public interface OAuthAccountRepository extends JpaRepository<OAuthAccount, Long> {
 
   Optional<OAuthAccount> findByProviderAndProviderUserId(String provider, String providerUserId);
 
-  Optional<OAuthAccount> findByMemberIdAndProvider(Integer memberId, String provider);
+  Optional<OAuthAccount> findByMemberIdAndProvider(Long memberId, String provider);
 
-  List<OAuthAccount> findAllByMemberIdOrderByIdAsc(Integer memberId);
+  List<OAuthAccount> findAllByMemberIdOrderByIdAsc(Long memberId);
 }
