+++ back/src/main/java/com/back/member/adapter/in/web/MemberController.java
@@ -34,7 +34,7 @@ public class MemberController {
   /** 본인 회원 정보 조회 엔드포인트. */
   @GetMapping("/me")
   public RsData<MemberResponse> getMyMember(Authentication authentication) {
-    Integer memberId = resolveAuthenticatedMemberId(authentication);
+    Long memberId = resolveAuthenticatedMemberId(authentication);
     return new RsData<>("200-1", "Member fetched.", memberService.getMember(memberId));
   }
 
@@ -42,7 +42,7 @@ public class MemberController {
   @Deprecated
   @PreAuthorize("hasRole('ADMIN')")
   @GetMapping("/{memberId}")
-  public RsData<MemberResponse> getMember(@PathVariable Integer memberId) {
+  public RsData<MemberResponse> getMember(@PathVariable Long memberId) {
     return new RsData<>("200-1", "Member fetched.", memberService.getMember(memberId));
   }
 
@@ -50,7 +50,7 @@ public class MemberController {
   @PatchMapping("/me/profile")
   public RsData<MemberResponse> updateMyProfile(
       Authentication authentication, @RequestBody UpdateMemberProfileRequest request) {
-    Integer memberId = resolveAuthenticatedMemberId(authentication);
+    Long memberId = resolveAuthenticatedMemberId(authentication);
     return new RsData<>(
         "200-2", "Member profile updated.", memberService.updateProfile(memberId, request));
   }
@@ -60,12 +60,12 @@ public class MemberController {
   @PreAuthorize("hasRole('ADMIN')")
   @PatchMapping("/{memberId}/profile")
   public RsData<MemberResponse> updateProfile(
-      @PathVariable Integer memberId, @RequestBody UpdateMemberProfileRequest request) {
+      @PathVariable Long memberId, @RequestBody UpdateMemberProfileRequest request) {
     return new RsData<>(
         "200-2", "Member profile updated.", memberService.updateProfile(memberId, request));
   }
 
-  private Integer resolveAuthenticatedMemberId(Authentication authentication) {
+  private Long resolveAuthenticatedMemberId(Authentication authentication) {
     if (authentication == null
         || !(authentication.getPrincipal() instanceof AuthenticatedMember principal)) {
       throw new ServiceException("401-1", "Authentication is required.");
